Refactored fake-controller to service
